don't use enum class as it makes the enum type too strong for now, go back to an old enum so it can be implicitly converted to uint32

+++ decompiler/simple_disassembler.h
@@ -75,9 +75,10 @@ public:
 		INC_ADDR; \
 	}
 
+
 #define OPCODE_BASE(val) \
-	case (uint32)val: \
-		full_opcode = (full_opcode << 8) + static_cast<uint32>(val);
+	case val: \
+		full_opcode = (full_opcode << 8) + val;
 
 #define OPCODE_END break;
